requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
googlesearch-python==1.2.4
idna==3.7
requests==2.32.3
soupsieve==2.5
tqdm==4.66.4
urllib3==2.2.2
